Fix transient dryer door detection
The dryer can (of course) transiently show power in the door-open
range when starting a wrinkle protect, just as it can when finishing a
cycle. Handle both transients.

Unfortunately, this makes it more sensitive to only refreshing when
there's a power update. Since it now takes a window to respond to the
door being open, but the door-open power is quite consistent, it
generally fails to respond to a door-open until the next 30 second
update from the power meter.